requirements.txt
Flask
psutil
supabase
python-dotenv
requests
gunicorn
flask-cors